// frontend/src/types/order/ShippingDetails.d.ts
export interface ShippingDetails {
    shippingName:       string;
    shippingAddress:    string;
    shippingCity:       string;
    shippingState:      string;
    shippingPostalCode: string;
    shippingCountry:    string;
    billingName:        string;
    billingAddress:     string;
    billingCity:        string;
    billingState:       string;
    billingPostalCode:  string;
    billingCountry:     string;
}